perf: Benchmark SiblingSubgraph construction
